Kindle to Anki converstion
Create library and script to query database on an E-Kindle device or from a file. If the definition is unknown, then it
is looked up using the database then the Merriam Webster Dictionary API. The word and its definition are then pushed to
the DB.

The words stored in the DB can be published to an Anki compatible file, which can be then ingested by the Anki program.